Got Response 200
I was fiddling around with it and slowly losing my sanity in the process, but I somehow managed to get it to a Response 200.
I know you wanted me to try something else, and I tried a lot of differennt things but I got it "working" with a different api.